+++ apps/storybook/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @pikas-ui/storybook
 
+## 1.0.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @pikas-ui/button@1.0.8
+  - @pikas-ui/icons@1.0.8
+  - @pikas-ui/loader@1.0.8
+  - @pikas-ui/styles@1.0.8
+
 ## 1.0.7
 
 ### Patch Changes