+++ packages/loader/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pikas-ui/loader
 
+## 1.0.8
+
+### Patch Changes
+
+- Add button library and icons library
+- Updated dependencies
+  - @pikas-ui/styles@1.0.8
+
 ## 1.0.7
 
 ### Patch Changes